refactor: extract evmTools into multiple files
