+++ src/toolcalls/evmTools/contractTools.ts
@@ -1,0 +1,211 @@
+import { z } from 'zod';
+import {
+  decodeFunctionResult,
+  encodeFunctionData,
+  type Abi,
+  type Hex,
+} from 'viem';
+import {
+  EvmToolOperation,
+  createToolName,
+  convertBigIntToString,
+} from './types';
+import { getEthereumProvider } from './helpers';
+
+// Call Contract Tool
+export class CallContractTool extends EvmToolOperation {
+  name = createToolName('call-contract');
+  description =
+    'Call a contract function. Use this to run smart contract methods.';
+  zodSchema = z.object({
+    to: z.string().describe('The contract address or recipient'),
+    data: z
+      .string()
+      .describe('A contract hashed method call with encoded args'),
+    value: z
+      .string()
+      .optional()
+      .describe('Value (in wei) sent with this transaction'),
+  });
+
+  async execute(params: unknown): Promise<string> {
+    const { to, data, value } = this.zodSchema.parse(params) as {
+      to: string;
+      data: string;
+      value?: string;
+    };
+
+    const provider = getEthereumProvider();
+    const callParameters = {
+      to,
+      data,
+      ...(value && { value }),
+    };
+
+    const result = (await provider.request({
+      method: 'eth_call',
+      params: [callParameters, 'latest'],
+    })) as string;
+
+    return result;
+  }
+}
+
+// Encode Function Data Tool
+export class EncodeFunctionDataTool extends EvmToolOperation {
+  name = createToolName('encode-function-data');
+  description =
+    'Encode function data for contract calls using ABI and parameters. Use this to prepare data for contract interactions.';
+  zodSchema = z.object({
+    abi: z.array(z.unknown()).describe('The contract ABI'),
+    functionName: z.string().describe('The function name to encode'),
+    args: z
+      .array(z.unknown())
+      .optional()
+      .describe('Arguments to pass to the function'),
+  });
+
+  async execute(params: unknown): Promise<string> {
+    const { abi, functionName, args } = this.zodSchema.parse(params) as {
+      abi: unknown[];
+      functionName: string;
+      args?: unknown[];
+    };
+
+    try {
+      const encodedData = encodeFunctionData({
+        abi: abi as Abi,
+        functionName,
+        args: args || [],
+      });
+
+      return encodedData;
+    } catch (error) {
+      throw new Error(
+        `Failed to encode function data: ${error instanceof Error ? error.message : 'Unknown error'}`,
+      );
+    }
+  }
+}
+
+// Estimate Gas Tool
+export class EstimateGasTool extends EvmToolOperation {
+  name = createToolName('estimate-gas');
+  description =
+    'Estimate the gas necessary to complete a transaction without submitting it to the network. Use this before sending transactions to estimate costs.';
+  zodSchema = z.object({
+    to: z.string().describe('The transaction recipient or contract address'),
+    data: z
+      .string()
+      .optional()
+      .describe('A contract hashed method call with encoded args'),
+    value: z
+      .string()
+      .optional()
+      .describe('Value in wei sent with this transaction'),
+    maxFeePerGas: z
+      .string()
+      .optional()
+      .describe('Total fee per gas in wei, inclusive of maxPriorityFeePerGas'),
+    maxPriorityFeePerGas: z
+      .string()
+      .optional()
+      .describe('Max priority fee per gas in wei'),
+    chainId: z
+      .number()
+      .optional()
+      .describe('Chain ID to validate against before sending transaction'),
+  });
+
+  async execute(params: unknown): Promise<string> {
+    const { to, data, value, maxFeePerGas, maxPriorityFeePerGas } =
+      this.zodSchema.parse(params) as {
+        to: string;
+        data?: string;
+        value?: string;
+        maxFeePerGas?: string;
+        maxPriorityFeePerGas?: string;
+        chainId?: number;
+      };
+
+    const provider = getEthereumProvider();
+    const estimateParameters = {
+      to,
+      ...(data && { data }),
+      ...(value && { value }),
+      ...(maxFeePerGas && { maxFeePerGas }),
+      ...(maxPriorityFeePerGas && { maxPriorityFeePerGas }),
+    };
+
+    const gasEstimate = (await provider.request({
+      method: 'eth_estimateGas',
+      params: [estimateParameters],
+    })) as string;
+
+    return gasEstimate;
+  }
+}
+
+// Read Contract Tool
+export class ReadContractTool extends EvmToolOperation {
+  name = createToolName('read-contract');
+  description =
+    'Call a read-only function on a contract, and return the response. Use this for reading any contract state with a specific ABI and function.';
+  zodSchema = z.object({
+    abi: z.array(z.unknown()).describe("The contract's ABI"),
+    address: z.string().describe("The contract's address"),
+    functionName: z.string().describe('Function to call on the contract'),
+    args: z
+      .array(z.unknown())
+      .optional()
+      .describe('Arguments to pass when calling the contract'),
+    chainId: z
+      .number()
+      .optional()
+      .describe('ID of chain to use when fetching data'),
+  });
+
+  async execute(params: unknown): Promise<string> {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const { abi, address, functionName, args, chainId } = this.zodSchema.parse(
+      params,
+    ) as {
+      abi: unknown[];
+      address: string;
+      functionName: string;
+      args?: unknown[];
+      chainId?: number;
+    };
+
+    try {
+      // Encode the function call
+      const data = encodeFunctionData({
+        abi: abi as Abi,
+        functionName,
+        args: args || [],
+      });
+
+      // Make the call
+      const provider = getEthereumProvider();
+      const result = (await provider.request({
+        method: 'eth_call',
+        params: [{ to: address, data }, 'latest'],
+      })) as string;
+
+      // Decode the result
+      const decodedResult = decodeFunctionResult({
+        abi: abi as Abi,
+        functionName,
+        data: result as Hex,
+      });
+
+      // Convert BigInt to string for JSON serialization
+      const serializableResult = convertBigIntToString(decodedResult);
+      return JSON.stringify(serializableResult);
+    } catch (error) {
+      throw new Error(
+        `Failed to read contract: ${error instanceof Error ? error.message : 'Unknown error'}`,
+      );
+    }
+  }
+}
