+++ src/toolcalls/evmTools/helpers.ts
@@ -1,0 +1,111 @@
+import { getAddress, isAddress } from 'viem';
+import { chainRegistry, EVMChainConfig } from '../chain/chainsRegistry';
+import { ChainType } from '../chain/constants';
+import { EthereumProvider } from './types';
+
+// Helper to get Ethereum provider
+export const getEthereumProvider = (): EthereumProvider => {
+  if (typeof window === 'undefined' || !window.ethereum) {
+    throw new Error(
+      'Ethereum provider not detected. Please install a compatible wallet extension and ensure it is connected.',
+    );
+  }
+  return window.ethereum;
+};
+
+// Helper to get current account
+export const getCurrentAccount = async (): Promise<string> => {
+  const provider = getEthereumProvider();
+  const accounts = (await provider.request({
+    method: 'eth_accounts',
+  })) as string[];
+  if (!accounts || accounts.length === 0) {
+    throw new Error('No account connected. Please connect your wallet.');
+  }
+  return accounts[0];
+};
+
+// Helper to get chain ID
+export const getCurrentChainId = async (): Promise<number> => {
+  const provider = getEthereumProvider();
+  const chainId = (await provider.request({
+    method: 'eth_chainId',
+  })) as string;
+  return parseInt(chainId, 16);
+};
+
+// Helper to get contract address for a token on current chain
+export const getContractAddress = async (
+  tokenSymbol: string,
+): Promise<string | null> => {
+  const chainId = await getCurrentChainId();
+  const chainInfo = getChainConfigByChainId(chainId.toString());
+
+  if (!chainInfo) {
+    return null;
+  }
+
+  const contract = chainInfo.chainConfig.erc20Contracts[tokenSymbol];
+  return contract?.contractAddress || null;
+};
+
+// Helper to get chain config for current chain
+export const getCurrentChainConfig = async () => {
+  const chainId = await getCurrentChainId();
+  const chainInfo = getChainConfigByChainId(chainId.toString());
+
+  if (!chainInfo) {
+    throw new Error(`Chain with ID ${chainId} not found in registry`);
+  }
+
+  return chainInfo;
+};
+
+// Helper to get chain config by chain ID for EVM chains
+export const getChainConfigByChainId = (
+  chainId: string,
+): { chainName: string; chainConfig: EVMChainConfig } | null => {
+  const evmChains = chainRegistry[ChainType.EVM];
+  for (const [chainName, chainConfig] of Object.entries(evmChains)) {
+    if (
+      chainConfig.chainType === ChainType.EVM &&
+      chainConfig.chainID === chainId
+    ) {
+      return {
+        chainName,
+        chainConfig: chainConfig as EVMChainConfig,
+      };
+    }
+  }
+  return null;
+};
+
+// Helper to get chain config by name for EVM chains
+export const getEvmChainConfigByName = (
+  chainName: string,
+): EVMChainConfig | null => {
+  const evmChains = chainRegistry[ChainType.EVM];
+  const chainConfig = evmChains[chainName];
+
+  if (chainConfig && chainConfig.chainType === ChainType.EVM) {
+    return chainConfig as EVMChainConfig;
+  }
+
+  return null;
+};
+
+// Helper to validate and normalize address
+export const validateAddress = (address: string): string => {
+  if (!isAddress(address)) {
+    throw new Error(`Invalid address: ${address}`);
+  }
+  return getAddress(address);
+};
+
+// Helper to validate and normalize contract address
+export const validateContractAddress = (symbolOrAddress: string): string => {
+  if (isAddress(symbolOrAddress)) {
+    return getAddress(symbolOrAddress);
+  }
+  throw new Error(`Invalid contract address: ${symbolOrAddress}`);
+};
